+++ packages/core/src/deployment.ts
@@ -10,12 +10,25 @@ import {
   isDevelopmentNetwork,
   isReceiptSuccessful,
 } from './provider';
+import { UpgradesError } from './error';
 
 const sleep = promisify(setTimeout);
 
 export interface Deployment {
   address: string;
   txHash?: string;
+}
+
+export interface DeployOpts {
+  /**
+   * Timeout in milliseconds to wait for the transaction confirmation when deploying an implementation contract or proxy admin contract. Use `0` to wait indefinitely.
+   */
+  timeout?: number;
+
+  /**
+   * Polling interval in milliseconds between checks for the transaction confirmation when deploying an implementation contract or proxy admin contract.
+   */
+  pollingInterval?: number;
 }
 
 export async function resumeOrDeploy<T extends Deployment>(
@@ -51,24 +64,25 @@ export async function resumeOrDeploy<T extends Deployment>(
   return deployment;
 }
 
-export async function waitAndValidateDeployment(provider: EthereumProvider, deployment: Deployment): Promise<void> {
+export async function waitAndValidateDeployment(
+  provider: EthereumProvider,
+  deployment: Deployment,
+  type?: string,
+  opts?: DeployOpts,
+): Promise<void> {
   const { txHash, address } = deployment;
 
-  // Poll for 60 seconds with a 5 second poll interval.
-  // TODO: Make these parameters configurable.
-  const pollTimeout = 60e3;
-  const pollInterval = 5e3;
+  // Poll for 60 seconds with a 5 second poll interval by default.
+  const pollTimeout = opts?.timeout ?? 60e3;
+  const pollInterval = opts?.pollingInterval ?? 5e3;
+
+  debug('polling timeout', pollTimeout, 'polling interval', pollInterval);
 
   if (txHash !== undefined) {
     const startTime = Date.now();
 
     // eslint-disable-next-line no-constant-condition
     while (true) {
-      const elapsedTime = Date.now() - startTime;
-      if (elapsedTime >= pollTimeout) {
-        // A timeout is NOT an InvalidDeployment
-        throw new TransactionMinedTimeout(deployment);
-      }
       debug('verifying deployment tx mined', txHash);
       const receipt = await getTransactionReceipt(provider, txHash);
       if (receipt && isReceiptSuccessful(receipt)) {
@@ -80,6 +94,13 @@ export async function waitAndValidateDeployment(provider: EthereumProvider, depl
       } else {
         debug('waiting for deployment tx mined', txHash);
         await sleep(pollInterval);
+      }
+      if (pollTimeout != 0) {
+        const elapsedTime = Date.now() - startTime;
+        if (elapsedTime >= pollTimeout) {
+          // A timeout is NOT an InvalidDeployment
+          throw new TransactionMinedTimeout(deployment, type, !!opts);
+        }
       }
     }
   }
@@ -96,9 +117,18 @@ export async function waitAndValidateDeployment(provider: EthereumProvider, depl
   debug('code in target address found', address);
 }
 
-export class TransactionMinedTimeout extends Error {
-  constructor(readonly deployment: Deployment) {
-    super(`Timed out waiting for transaction ${deployment.txHash}`);
+export class TransactionMinedTimeout extends UpgradesError {
+  constructor(readonly deployment: Deployment, type?: string, configurableTimeout?: boolean) {
+    super(
+      `Timed out waiting for ${type ? type + ' ' : ''}contract deployment to address ${
+        deployment.address
+      } with transaction ${deployment.txHash}`,
+      () =>
+        'Run the function again to continue waiting for the transaction confirmation.' +
+        (configurableTimeout
+          ? ' If the problem persists, adjust the polling parameters with the timeout and pollingInterval options.'
+          : ''),
+    );
   }
 }