+++ packages/plugin-hardhat/src/utils/deploy-impl.ts
@@ -108,11 +108,16 @@ async function deployImpl(
     assertStorageUpgradeSafe(currentLayout, deployData.layout, deployData.fullOpts);
   }
 
-  const impl = await fetchOrDeploy(deployData.version, deployData.provider, async () => {
-    const abi = ImplFactory.interface.format(FormatTypes.minimal) as string[];
-    const deployment = Object.assign({ abi }, await deploy(ImplFactory, ...deployData.fullOpts.constructorArgs));
-    return { ...deployment, layout };
-  });
+  const impl = await fetchOrDeploy(
+    deployData.version,
+    deployData.provider,
+    async () => {
+      const abi = ImplFactory.interface.format(FormatTypes.minimal) as string[];
+      const deployment = Object.assign({ abi }, await deploy(ImplFactory, ...deployData.fullOpts.constructorArgs));
+      return { ...deployment, layout };
+    },
+    opts,
+  );
 
   return { impl, kind: opts.kind };
 }